+++ VehEvalu8/Models/Motocross/Suzuki.cs
@@ -4,7 +4,7 @@ namespace TwoWheelTrader.Models.Motocross
 {
     public class Suzuki : IMotocross
     {
-        public Suzuki(int cc, int year, int priceForeign, string link, int distance)
+        public Suzuki(int cc, int year, double priceForeign, string link, int distance)
         {
             Make = "Suzuki";
             Model = "RM-Z";
